fix crash caused by previous commit
i found the reason why the call to delete was commented out.
so, one of the calls to `setDataSource` sets a `this` pointer
which should not be deleted.

as a work around, i added a bool to the function that when true, delete is used.
not the best fix in the world, but its not worth a proper fix as the whole api
should ditch raw new and malloc in favour of smart pointers anyway...

see #1